// src/features/personal-info/types.ts
export interface PersonalInfo {
  firstName: string;
  lastName: string;
  title: string;
  email: string;
  phone: string;
  city: string;
  github: string;
  linkedin: string;
  website: string;
}

export const initialPersonalInfo: PersonalInfo = {
  firstName: "",
  lastName: "",
  title: "",
  email: "",
  phone: "",
  city: "",
  github: "",
  linkedin: "",
  website: "",
};